adds option to override not installing camera settings
When the new revision of the AI cam is detected, the script shows an error,
and prevents installing the macros.
This makes sense if the AI cam is the only installed cam, but not if one has
installed an additional USB cam, connected to the front USB port.

The change gives an option, too continue the installation in that case,
by asking if the user wants to continue.