allows installing camera settings with additional USB cam
The old script didn't allow to install the camera settings macros, when the new HW revision of the AI cam has been detected,
even when the additional USB cam could benefit from being controlled by the macros.

The changes made in this commit give a warning about the new HW revision not bein compatible but allow the user to acknowledge
that he wants the camera settings macros installed nevertheless.